+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### changes
 
+- move `sine` utility function to `weldx.welding.util` [[#439]](https://github.com/BAMWelDX/weldx/pull/439)
+
 ### fixes
 
 - `WeldxFile.show_asdf_header` prints output on console, before it only returned the header